# springboot/src/main/java/bengen/bean/CreateCourseRequest.java
package bengen.bean;

import java.util.List;
import lombok.Getter;
import lombok.Setter;
import bengen.bean.*;


@Getter
@Setter
public class CreateCourseRequest {
    private Enumstatus status;
    private float price;
    private String title;
    private String cover;
    private int presenterRef;
    private String shortDesc;
    private int albumRef;
    private int homepageRef;
    private String poster;
    private CourseDto paramCourse;

}